fix ln invocation when lock link exists
When the lock symlink /var/run/zrep.lock already exists and we try
to create it again from another process (to acquire the lock), since
the target (e.g. /proc/1) is a directory, then ln tries to create
the symlink inside the already present target. Effectively, the
second process (e.g. 2) tries to create /var/run/zrep.lock/2 (where
zrep.lock resolves to /proc/1, so it's trying to create a symlink
/proc/1/2 pointing to /proc/2).

This makes no sense. The desired behaviour is to try (and fail) to
overwrite the link (/var/run/zrep.link) if it exists.

GNU and BSD variants have different switches for this behaviour,
hence the if.